+++ packages/google_sign_in/google_sign_in_platform_interface/lib/google_sign_in_platform_interface.dart
@@ -78,6 +78,9 @@ abstract class GoogleSignInPlatform {
   /// The [signInOption] determines the user experience. [SigninOption.games] is
   /// only supported on Android.
   ///
+  /// This method is deprecated in favor of [initWithForceFodeForRefreshToken] and
+  /// will be removed in a future update.
+  ///
   /// See:
   /// https://developers.google.com/identity/sign-in/web/reference#gapiauth2initparams
   Future<void> init({
@@ -85,9 +88,42 @@ abstract class GoogleSignInPlatform {
     SignInOption signInOption = SignInOption.standard,
     String? hostedDomain,
     String? clientId,
-    bool forceCodeForRefreshToken = false,
   }) async {
     throw UnimplementedError('init() has not been implemented.');
+  }
+
+  /// Initializes the plugin. You must call this method before calling other
+  /// methods.
+  ///
+  /// The [hostedDomain] argument specifies a hosted domain restriction. By
+  /// setting this, sign in will be restricted to accounts of the user in the
+  /// specified domain. By default, the list of accounts will not be restricted.
+  ///
+  /// The list of [scopes] are OAuth scope codes to request when signing in.
+  /// These scope codes will determine the level of data access that is granted
+  /// to your application by the user. The full list of available scopes can be
+  /// found here: <https://developers.google.com/identity/protocols/googlescopes>
+  ///
+  /// The [signInOption] determines the user experience. [SigninOption.games] is
+  /// only supported on Android.
+  ///
+  /// The [forceCodeForRefreshToken] is used on Android to ensure the authentication
+  /// code can be exchanged for a refresh token after the first request.
+  ///
+  /// See:
+  /// https://developers.google.com/identity/sign-in/web/reference#gapiauth2initparams
+  Future<void> initWithForceCodeForRefreshToken({
+    List<String> scopes = const <String>[],
+    SignInOption signInOption = SignInOption.standard,
+    String? hostedDomain,
+    String? clientId,
+    bool forceCodeForRefreshToken = false,
+  }) async {
+    await init(
+      scopes: scopes,
+      signInOption: signInOption,
+      hostedDomain: hostedDomain,
+      clientId: clientId);
   }
 
   /// Attempts to reuse pre-existing credentials to sign in again, without user interaction.