+++ packages/google_sign_in/google_sign_in/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 5.2.3
+
+* Add `forceCodeForRefreshToken` parameter to `GoogleSignIn`.
+
 ## 5.2.2
 
 * Updates Android compileSdkVersion to 31.